neural_claim_module/requirements.txt
pytesseract
Pillow
fastapi
uvicorn
requests
python-multipart
